third_party/blink/web_tests/external/wpt/css/css-align/blocks/justify-self-htb-rtl-vlr-expected.txt

This is a testharness.js-based test.
[FAIL] .item 4
  assert_equals: \n<div class="container">\n  <div class="item" style="justify-self: self-start;" data-expected-width="20" data-offset-x="0"></div>\n</div>\noffsetLeft expected 0 but got 20
[FAIL] .item 12
  assert_equals: \n<div class="container">\n  <div class="item rtl" style="justify-self: self-start;" data-expected-width="20" data-offset-x="0"></div>\n</div>\noffsetLeft expected 0 but got 20
Harness: the test ran to completion.
